unify conf.py and ad clean option to build.py

+++ build.py
@@ -6,9 +6,23 @@ import os
 
 langs = ['en', 'tr', 'zh-CN', 'hu', 'fr']
 
+clean = 0
+args = sys.argv[1:]
+if len(args) == 1:
+  if args[0] == "clean": clean = 1
+  else:
+    print("Unknown option. Usage: ./build.py [clean]") 
 
-cmd_clean = " echo 'Clean up...\n'; rm -r html xetex; "
-cmd_pdf =   " echo '------------\nBuild PDF...\n------------'; sphinx-build -b latex . xetex && cd xetex; xelatex *.tex; cp -f LittlevGL.pdf ../LittlevGL.pdf; cd ..; "
+
+if clean:
+  cmd_clean = " echo 'Clean up...\n'; rm -r html xetex; "
+else: 
+  cmd_clean = ""
+  
+print(cmd_clean)
+exit()
+  
+cmd_pdf =   " echo '------------\nBuild PDF...\n------------'; sphinx-build -b latex . xetex && cd xetex; xelatex -interaction=batchmode *.tex; xelatex -interaction=batchmode *.tex;  cp -f LittlevGL.pdf ../LittlevGL.pdf; cd ..; "
 cmd_html =  " echo '------------\nBuild HTML...\n------------'; sphinx-build -b html . html; "
 
 cmd_build = cmd_clean + cmd_pdf + cmd_html + " "
